Refactor the AuthClient inheritance tree
- introduce AuthLoginClient and AuthServiceClient as the dividing line
  between each part of the tree
- AuthClient becomes an alias for AuthLoginClient
- the "app client" classes inherit from login client
- a new base class, AuthBaseClient, provides OIDC and JWK methods,
  base data like service name, scopes, and error class, but no other
  methods

The calls provided on AuthBaseClient are those which are simple GET
methods (and a decorated version thereof) which do not require
authentication. As such, they make sense to expose equally to all
client types.
