+++ src/globus_sdk/_generate_init.py
@@ -110,11 +110,18 @@ _LAZY_IMPORT_TABLE: list[tuple[str, tuple[str, ...]]] = [
     (
         "services.auth",
         (
-            "AuthAPIError",
+            # client classes
+            "AuthBaseClient",
             "AuthClient",
-            "ConfidentialAppAuthClient",
-            "IdentityMap",
+            "AuthServiceClient",
+            "AuthLoginClient",
             "NativeAppAuthClient",
+            "ConfidentialAppAuthClient",
+            # errors
+            "AuthAPIError",
+            # high-level helpers
+            "IdentityMap",
+            # responses
             "GetIdentitiesResponse",
             "OAuthDependentTokenResponse",
             "OAuthTokenResponse",
