+++ src/globus_sdk/services/auth/client/__init__.py
@@ -1,5 +1,14 @@
-from .base import AuthClient
+from .base import AuthBaseClient
+from .base_login_client import AuthLoginClient
 from .confidential_client import ConfidentialAppAuthClient
 from .native_client import NativeAppAuthClient
+from .service_client import AuthClient, AuthServiceClient
 
-__all__ = ["AuthClient", "NativeAppAuthClient", "ConfidentialAppAuthClient"]
+__all__ = (
+    "AuthBaseClient",
+    "AuthClient",
+    "AuthServiceClient",
+    "AuthLoginClient",
+    "NativeAppAuthClient",
+    "ConfidentialAppAuthClient",
+)
