+++ tests/functional/services/auth/conftest.py
@@ -4,8 +4,16 @@ import globus_sdk
 
 
 @pytest.fixture
-def client(no_retry_transport):
-    class CustomAuthClient(globus_sdk.AuthClient):
+def login_client(no_retry_transport):
+    class CustomAuthClient(globus_sdk.AuthLoginClient):
+        transport_class = no_retry_transport
+
+    return CustomAuthClient()
+
+
+@pytest.fixture
+def service_client(no_retry_transport):
+    class CustomAuthClient(globus_sdk.AuthServiceClient):
         transport_class = no_retry_transport
 
     return CustomAuthClient()
