+++ tests/functional/services/auth/service_client/test_update_project.py
@@ -9,8 +9,8 @@ from globus_sdk._testing import get_last_request, load_response
 @pytest.mark.parametrize(
     "admin_id_style", ("none", "string", "list", "set", "uuid", "uuid_list")
 )
-def test_update_project_admin_id_styles(client, admin_id_style):
-    meta = load_response(client.update_project).metadata
+def test_update_project_admin_id_styles(service_client, admin_id_style):
+    meta = load_response(service_client.update_project).metadata
 
     if admin_id_style == "none":
         admin_ids = None
@@ -28,7 +28,7 @@ def test_update_project_admin_id_styles(client, admin_id_style):
         raise NotImplementedError(f"unknown admin_id_style {admin_id_style}")
 
     project_id = meta["id"]
-    res = client.update_project(
+    res = service_client.update_project(
         project_id, display_name="My Project", admin_ids=admin_ids
     )
 
@@ -49,8 +49,8 @@ def test_update_project_admin_id_styles(client, admin_id_style):
 @pytest.mark.parametrize(
     "group_id_style", ("string", "list", "set", "uuid", "uuid_list")
 )
-def test_update_project_group_id_styles(client, group_id_style):
-    meta = load_response(client.update_project, case="admin_group").metadata
+def test_update_project_group_id_styles(service_client, group_id_style):
+    meta = load_response(service_client.update_project, case="admin_group").metadata
 
     if group_id_style == "string":
         group_ids = meta["admin_group_id"]
@@ -66,7 +66,7 @@ def test_update_project_group_id_styles(client, group_id_style):
         raise NotImplementedError(f"unknown group_id_style {group_id_style}")
 
     project_id = meta["id"]
-    res = client.update_project(
+    res = service_client.update_project(
         project_id, contact_email="support@globus.org", admin_group_ids=group_ids
     )
