+++ tests/functional/services/auth/test_identity_map.py
@@ -28,17 +28,9 @@ IDENTITIES_MULTIPLE_RESPONSE = {
 }
 
 
-@pytest.fixture
-def client(no_retry_transport):
-    class CustomAuthClient(globus_sdk.AuthClient):
-        transport_class = no_retry_transport
-
-    return CustomAuthClient()
-
-
-def test_identity_map(client):
-    meta = load_response(client.get_identities, case="sirosen").metadata
-    idmap = globus_sdk.IdentityMap(client, [meta["username"]])
+def test_identity_map(service_client):
+    meta = load_response(service_client.get_identities, case="sirosen").metadata
+    idmap = globus_sdk.IdentityMap(service_client, [meta["username"]])
     assert idmap[meta["username"]]["organization"] == meta["org"]
 
     # lookup by ID also works
@@ -50,16 +42,16 @@ def test_identity_map(client):
     assert last_req.params == {"usernames": meta["username"], "provision": "false"}
 
 
-def test_identity_map_initialization_no_values(client):
-    idmap = globus_sdk.IdentityMap(client)
+def test_identity_map_initialization_no_values(service_client):
+    idmap = globus_sdk.IdentityMap(service_client)
     assert idmap.unresolved_ids == set()
     assert idmap.unresolved_usernames == set()
 
 
-def test_identity_map_initialization_mixed_and_duplicate_values(client):
+def test_identity_map_initialization_mixed_and_duplicate_values(service_client):
     # splits things up and deduplicates values into sets
     idmap = globus_sdk.IdentityMap(
-        client,
+        service_client,
         [
             "sirosen@globus.org",
             "ae341a98-d274-11e5-b888-dbae3a8ba545",
@@ -75,24 +67,24 @@ def test_identity_map_initialization_mixed_and_duplicate_values(client):
     assert idmap.unresolved_usernames == {"sirosen@globus.org", "globus@globus.org"}
 
 
-def test_identity_map_initialization_batch_size(client):
-    idmap = globus_sdk.IdentityMap(client, id_batch_size=10)
+def test_identity_map_initialization_batch_size(service_client):
+    idmap = globus_sdk.IdentityMap(service_client, id_batch_size=10)
     assert idmap.unresolved_ids == set()
     assert idmap.unresolved_usernames == set()
     assert idmap.id_batch_size == 10
 
 
-def test_identity_map_add(client):
-    idmap = globus_sdk.IdentityMap(client)
+def test_identity_map_add(service_client):
+    idmap = globus_sdk.IdentityMap(service_client)
     assert idmap.add("sirosen@globus.org") is True
     assert idmap.add("sirosen@globus.org") is False
     assert idmap.add("46bd0f56-e24f-11e5-a510-131bef46955c") is True
     assert idmap.add("46bd0f56-e24f-11e5-a510-131bef46955c") is False
 
 
-def test_identity_map_add_after_lookup(client):
-    meta = load_response(client.get_identities, case="sirosen").metadata
-    idmap = globus_sdk.IdentityMap(client)
+def test_identity_map_add_after_lookup(service_client):
+    meta = load_response(service_client.get_identities, case="sirosen").metadata
+    idmap = globus_sdk.IdentityMap(service_client)
     x = idmap[meta["username"]]["id"]
     # this is the key: adding it will indicate that we've already seen this ID, perhaps
     # "unintuitively", and that's part of the value of `add()` returning a boolean value
@@ -100,9 +92,11 @@ def test_identity_map_add_after_lookup(client):
     assert idmap[x] == idmap[meta["username"]]
 
 
-def test_identity_map_multiple(client):
-    meta = load_response(client.get_identities, case="multiple").metadata
-    idmap = globus_sdk.IdentityMap(client, ["sirosen@globus.org", "globus@globus.org"])
+def test_identity_map_multiple(service_client):
+    meta = load_response(service_client.get_identities, case="multiple").metadata
+    idmap = globus_sdk.IdentityMap(
+        service_client, ["sirosen@globus.org", "globus@globus.org"]
+    )
     assert idmap["sirosen@globus.org"]["organization"] == "Globus Team"
     assert idmap["globus@globus.org"]["organization"] is None
 
@@ -117,9 +111,9 @@ def test_identity_map_multiple(client):
     assert last_req.params["provision"] == "false"
 
 
-def test_identity_map_keyerror(client):
-    load_response(client.get_identities, case="sirosen")
-    idmap = globus_sdk.IdentityMap(client)
+def test_identity_map_keyerror(service_client):
+    load_response(service_client.get_identities, case="sirosen")
+    idmap = globus_sdk.IdentityMap(service_client)
     # a name which doesn't come back, indicating that it was not found, will KeyError
     with pytest.raises(KeyError):
         idmap["sirosen2@globus.org"]
@@ -128,18 +122,18 @@ def test_identity_map_keyerror(client):
     assert last_req.params == {"usernames": "sirosen2@globus.org", "provision": "false"}
 
 
-def test_identity_map_get_with_default(client):
-    load_response(client.get_identities, case="sirosen")
+def test_identity_map_get_with_default(service_client):
+    load_response(service_client.get_identities, case="sirosen")
     magic = object()  # sentinel value
-    idmap = globus_sdk.IdentityMap(client)
+    idmap = globus_sdk.IdentityMap(service_client)
     # a name which doesn't come back, if looked up with `get()` should return the
     # default
     assert idmap.get("sirosen2@globus.org", magic) is magic
 
 
-def test_identity_map_del(client):
-    meta = load_response(client.get_identities).metadata
-    idmap = globus_sdk.IdentityMap(client)
+def test_identity_map_del(service_client):
+    meta = load_response(service_client.get_identities).metadata
+    idmap = globus_sdk.IdentityMap(service_client)
     identity_id = idmap[meta["username"]]["id"]
     del idmap[identity_id]
     assert idmap.get(meta["username"])["id"] == identity_id
@@ -160,12 +154,12 @@ def test_identity_map_del(client):
     ],
 )
 @pytest.mark.parametrize("initial_add", [True, False])
-def test_identity_map_shared_cache_match(client, initial_add, lookup1, lookup2):
-    meta = load_response(client.get_identities, case="sirosen").metadata
+def test_identity_map_shared_cache_match(service_client, initial_add, lookup1, lookup2):
+    meta = load_response(service_client.get_identities, case="sirosen").metadata
     lookup1, lookup2 = meta[lookup1], meta[lookup2]
     cache = {}
-    idmap1 = globus_sdk.IdentityMap(client, cache=cache)
-    idmap2 = globus_sdk.IdentityMap(client, cache=cache)
+    idmap1 = globus_sdk.IdentityMap(service_client, cache=cache)
+    idmap2 = globus_sdk.IdentityMap(service_client, cache=cache)
     if initial_add:
         idmap1.add(lookup1)
         idmap2.add(lookup2)
@@ -187,12 +181,12 @@ def test_identity_map_shared_cache_match(client, initial_add, lookup1, lookup2):
     "lookup_style",
     ["usernames", "ids"],
 )
-def test_identity_map_shared_cache_mismatch(client, lookup_style):
-    meta = load_response(client.get_identities, case="multiple").metadata
+def test_identity_map_shared_cache_mismatch(service_client, lookup_style):
+    meta = load_response(service_client.get_identities, case="multiple").metadata
     lookup1, lookup2 = meta[lookup_style]
     cache = {}
-    idmap1 = globus_sdk.IdentityMap(client, [lookup1, lookup2], cache=cache)
-    idmap2 = globus_sdk.IdentityMap(client, cache=cache)
+    idmap1 = globus_sdk.IdentityMap(service_client, [lookup1, lookup2], cache=cache)
+    idmap2 = globus_sdk.IdentityMap(service_client, cache=cache)
     # no requests yet...
     assert len(responses.calls) == 0
     # do the first lookup, it should make one request
@@ -207,12 +201,12 @@ def test_identity_map_shared_cache_mismatch(client, lookup_style):
     assert len(responses.calls) == 1
 
 
-def test_identity_map_prepopulated_cache(client):
-    meta = load_response(client.get_identities).metadata
+def test_identity_map_prepopulated_cache(service_client):
+    meta = load_response(service_client.get_identities).metadata
 
     # populate the cache, even with nulls it should stop any lookups from happening
     cache = {meta["id"]: None, meta["username"]: None}
-    idmap = globus_sdk.IdentityMap(client, cache=cache)
+    idmap = globus_sdk.IdentityMap(service_client, cache=cache)
     # no requests yet...
     assert len(responses.calls) == 0
     # do the lookups
@@ -222,12 +216,12 @@ def test_identity_map_prepopulated_cache(client):
     assert len(responses.calls) == 0
 
 
-def test_identity_map_batch_limit(client):
-    meta1 = load_response(client.get_identities).metadata
-    meta2 = load_response(client.get_identities, case="sirosen").metadata
+def test_identity_map_batch_limit(service_client):
+    meta1 = load_response(service_client.get_identities).metadata
+    meta2 = load_response(service_client.get_identities, case="sirosen").metadata
 
     # setup the ID map with a size limit of 1
-    idmap = globus_sdk.IdentityMap(client, id_batch_size=1)
+    idmap = globus_sdk.IdentityMap(service_client, id_batch_size=1)
     idmap.add(meta2["id"])
     idmap.add(meta1["id"])
