+++ tests/non-pytest/mypy-ignore-tests/get_authorize_url_supports_session_params.py
@@ -3,7 +3,7 @@ import uuid
 
 import globus_sdk
 
-ac = globus_sdk.AuthClient()
+ac = globus_sdk.AuthLoginClient("someclientid")
 
 # base case
 url: str = ac.oauth2_get_authorize_url()
