+++ tests/unit/test_auth_clients.py
@@ -3,28 +3,60 @@ import uuid
 
 import pytest
 
-from globus_sdk import AuthClient, ConfidentialAppAuthClient, NativeAppAuthClient
+import globus_sdk
 
 CLIENT_ID_UUID = uuid.uuid4()
 CLIENT_ID_STR = str(CLIENT_ID_UUID)
 
 
-def test_base_auth_client_does_not_require_client_id():
-    client = AuthClient()
-    assert client.client_id is None
+def test_service_client_does_not_require_client_id():
+    client = globus_sdk.AuthServiceClient()
+    # accessing the attribute warns, but provides None
+    with pytest.warns(globus_sdk.RemovedInV4Warning):
+        assert client.client_id is None
+
+
+@pytest.mark.parametrize("pass_value_as", ("str", "uuid"))
+def test_service_client_allows_client_id_but_warns(pass_value_as):
+    pass_value = CLIENT_ID_UUID if pass_value_as == "uuid" else CLIENT_ID_STR
+
+    # init will warn because a value is being passed
+    with pytest.warns(globus_sdk.RemovedInV4Warning):
+        client = globus_sdk.AuthServiceClient(client_id=pass_value)
+
+    # accessing the attribute warns a second time, but provides the stringified value
+    with pytest.warns(globus_sdk.RemovedInV4Warning):
+        assert client.client_id == CLIENT_ID_STR
+
+
+@pytest.mark.parametrize("pass_value_as", ("str", "uuid"))
+def test_service_client_allows_client_id_assignment(pass_value_as):
+    pass_value = CLIENT_ID_UUID if pass_value_as == "uuid" else CLIENT_ID_STR
+
+    client = globus_sdk.AuthServiceClient()
+    with pytest.warns(globus_sdk.RemovedInV4Warning):
+        client.client_id = pass_value
+
+    with pytest.warns(globus_sdk.RemovedInV4Warning):
+        assert client.client_id == CLIENT_ID_STR
 
 
 @pytest.mark.parametrize(
-    "client_type", (AuthClient, ConfidentialAppAuthClient, NativeAppAuthClient)
+    "client_type",
+    (
+        globus_sdk.AuthLoginClient,
+        globus_sdk.ConfidentialAppAuthClient,
+        globus_sdk.NativeAppAuthClient,
+    ),
 )
 @pytest.mark.parametrize("pass_value_as", ("str", "uuid"))
 def test_can_use_uuid_or_str_for_client_id(client_type, pass_value_as):
     pass_value = CLIENT_ID_UUID if pass_value_as == "uuid" else CLIENT_ID_STR
 
-    if client_type in (AuthClient, NativeAppAuthClient):
+    if client_type in (globus_sdk.AuthLoginClient, globus_sdk.NativeAppAuthClient):
         client = client_type(client_id=pass_value)
-    elif client_type is ConfidentialAppAuthClient:
-        client = ConfidentialAppAuthClient(pass_value, "bogus_secret")
+    elif client_type is globus_sdk.ConfidentialAppAuthClient:
+        client = globus_sdk.ConfidentialAppAuthClient(pass_value, "bogus_secret")
     else:
         raise NotImplementedError
 
@@ -34,10 +66,12 @@ def test_can_use_uuid_or_str_for_client_id(client_type, pass_value_as):
 def test_native_app_auth_client_rejects_authorizer():
     authorizer = unittest.mock.Mock()
     with pytest.raises(TypeError):
-        NativeAppAuthClient(CLIENT_ID_UUID, authorizer=authorizer)
+        globus_sdk.NativeAppAuthClient(CLIENT_ID_UUID, authorizer=authorizer)
 
 
 def test_confidential_app_auth_client_rejects_authorizer():
     authorizer = unittest.mock.Mock()
     with pytest.raises(TypeError):
-        ConfidentialAppAuthClient(CLIENT_ID_UUID, "foo-secret", authorizer=authorizer)
+        globus_sdk.ConfidentialAppAuthClient(
+            CLIENT_ID_UUID, "foo-secret", authorizer=authorizer
+        )
